Completely reimplement using protocol buffer
So it turns out the i6 protocol uses protobuf with a simple framing
scheme to put one or more serialized protobuf on a stream connection.
The previous manual parsing code was just poorly re-implementation a
protobuf deserializer.

This patch throws that away and replaces all that code with a nice proto
file.

This patch also fixes embarrassing Python syntax mistakes in the emulation
prevention removal code that was corrupting certain messages (the code
was simply not working). To avoid this in the future, the patch also
adds tests for that function.