Reimplement connection and query loops with Protocol and run loop
This patch reimplements the core connection and query loops using a
really simple Protocol and just run loop APIs. This improves performance
by avoiding the creation of tasks and coroutines.

This patch also makes device availability more robust by looking for a
core set of so-called required properties to be set in the internal
Properties protocol buffer, instead of using a "idle time since last
query results", which was prone to misfire.