(feat) Suppress client callbacks on redundant changes
This patch changes the behavior of aiobafi6 to suppress client callbacks
when no property value has changed when processing query results sent
by a device.

This patch also introduces a list of volatile properties that tend to
change frequently or erratically (such as fan RPM and uptime). These
properties are by-default ignored when determining if callbacks should
be invoked. A device init setting can disable this behavior.