Detect service and device UUID mismatches (#72)
* Detect service and device UUID mismatches

This patch introduces a new validation stage to the Device connection
flow where the UUID in the Service object used to initialize the device
is validated against the UUID received in the initial device properties
query. If the UUIDs do not match, the connection is closed and an
exception is set on the run future and any "wait for available"
coroutines.

This feature required a fairly big rework of internal device connection
logic, but the basic design remains the same. Additional tests cover the
new functionality.

See https://github.com/home-assistant/core/pull/98784

* Update aiobafi6/device.py

Co-authored-by: J. Nick Koston <nick@koston.org>

* Update aiobafi6/device.py

Co-authored-by: J. Nick Koston <nick@koston.org>

---------

Co-authored-by: J. Nick Koston <nick@koston.org>